JW-VarietyPrice: 품질별 가격 차트 구현

+++ backend/core/src/main/java/com/bombi/core/presentation/dto/price/QualityChartData.java
@@ -9,12 +9,12 @@ public class QualityChartData {
 
 	private String date;
 	private String variety;
-	private int special;
-	private int high;
-	private int moderate;
-	private int other;
+	private Long special;
+	private Long high;
+	private Long moderate;
+	private Long other;
 
-	public QualityChartData(String date, String variety, int special, int high, int moderate, int other) {
+	public QualityChartData(String date, String variety, Long special, Long high, Long moderate, Long other) {
 		this.date = date;
 		this.variety = variety;
 		this.special = special;
